+++ projects/ng-translation/src/lib/models/resource.model.ts
@@ -1,0 +1,12 @@
+/* 3rd party libraries */
+
+/* locally accessible feature module code, always use relative path */
+
+export interface Resource {
+
+  name: string;
+  alias: string;
+  url: string;
+  format: string;
+  inUse: boolean;
+}